+++ lib/src/model/ewelink_error_response.dart
@@ -20,17 +20,32 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
 
-import 'ewelink_credentials.dart';
+import 'package:json_annotation/json_annotation.dart';
 
-class Ewelink {
-  Ewelink({this.email, this.password, required this.region, this.credentials});
-  String? email;
-  String? password;
-  String region;
-  EwelinkCredentials? credentials;
+import 'ewelink_user.dart';
 
-  factory Ewelink.fromCredentilals({
-    required EwelinkCredentials credentials,
-  }) =>
-      Ewelink(credentials: credentials, region: credentials.region);
+part 'ewelink_error_response.g.dart';
+
+@JsonSerializable(explicitToJson: true)
+// The api always returns status code 200, even when there are errors.
+// If has erros, then the error will be inside the body.
+class EwelinkErrorResponse {
+  EwelinkErrorResponse({
+    this.msg = '',
+    this.errmsg = '',
+    required this.code,
+  });
+  String msg;
+  String errmsg;
+  @JsonKey(name: 'error')
+  int code;
+
+  factory EwelinkErrorResponse.fromJson(Map<String, dynamic> json) =>
+      _$EwelinkErrorResponseFromJson(json);
+  Map<String, dynamic> toJson() => _$EwelinkErrorResponseToJson(this);
+
+  static bool hasError(Map<String, dynamic> response) {
+    int error = (response['error'] as int?) ?? 0;
+    return (error != 0);
+  }
 }