+++ CHANGELOG.md
@@ -1,3 +1,6 @@
 ## 0.0.1
 
-* TODO: Describe initial release.
+* Implement toogleDevice() with hardcoded initial state.
+* Implement getCredentials() from email and password.
+* Other methods implemented to allow toggleDevice usage:
+  setDevicePowerState() and getDevice()